+++ src/main/java/roomescape/service/ThemeService.java
@@ -4,8 +4,11 @@ import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
+import roomescape.exception.CustomException;
+import roomescape.exception.ErrorCode;
 import roomescape.service.dto.ServiceThemeRequest;
 import roomescape.service.dto.ServiceThemeResponse;
 
@@ -13,11 +16,15 @@ import roomescape.service.dto.ServiceThemeResponse;
 @Transactional(readOnly = true)
 public class ThemeService {
     public static final int RANKING_LIMIT = 10;
-    private final ThemeDao themeDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
+
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
     }
+
 
     @Transactional
     public ServiceThemeResponse create(ServiceThemeRequest requestDto) {
@@ -33,6 +40,9 @@ public class ThemeService {
 
     @Transactional
     public void delete(Long id) {
+        if (reservationDao.existByThemeId(id)) {
+            throw new CustomException(ErrorCode.REFERENCED_THEME);
+        }
         themeDao.delete(id);
     }
